+++ src/main/java/one/vladimir/wts/DBModule/Repositories/GroupRepository.java
@@ -3,5 +3,5 @@ package one.vladimir.wts.DBModule.Repositories;
 import one.vladimir.wts.DBModule.Entities.Group;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GroupRepository extends CrudRepository<Group, Long> {
+public interface GroupRepository extends CrudRepository<Group, Integer> {
 }
