+++ src/main/java/one/vladimir/wts/DBModule/Repositories/PointRepository.java
@@ -3,5 +3,5 @@ package one.vladimir.wts.DBModule.Repositories;
 import one.vladimir.wts.DBModule.Entities.Point;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PointRepository extends CrudRepository<Point, Long> {
+public interface PointRepository extends CrudRepository<Point, Integer> {
 }
